+++ src/core/index.ts
@@ -13,4 +13,5 @@ export * from "./services/query.service.js";
 export * from "./services/reactive.service.js";
 export * from "./services/timing.service.js";
 export * from "./services/utils.service.js";
+export * from "./services/validation.service.js";
 export * from "./services/youtube.service.js";